do test json review at end to properly fail job
