+++ test/new-e2e/system-probe/test-runner/main.go
@@ -6,9 +6,6 @@
 package main
 
 import (
-	"bufio"
-	"bytes"
-	"encoding/json"
 	"errors"
 	"fmt"
 	"io"
@@ -128,67 +125,36 @@ func mergeEnv(env ...map[string]interface{}) []string {
 	return mergedEnv
 }
 
-type testEvent struct {
-	Time    time.Time // encodes as an RFC3339-format string
-	Action  string
-	Package string
-	Test    string
-	Elapsed float64 // seconds
-	Output  string
-}
-
 // concatenateJsons combines all the test json output files into a single file.
-// It also returns a formatted string containing all the failed tests.
-func concatenateJsons(indir, outdir string) (string, error) {
+func concatenateJsons(indir, outdir string) error {
 	testJsonFile := filepath.Join(outdir, "out.json")
 	matches, err := glob(indir, `.*\.json`, func(path string) bool { return true })
 	if err != nil {
-		return "", fmt.Errorf("json glob: %s", err)
+		return fmt.Errorf("json glob: %s", err)
 	}
 
 	f, err := os.OpenFile(testJsonFile, os.O_APPEND|os.O_CREATE|os.O_WRONLY, 0666)
 	if err != nil {
-		return "", fmt.Errorf("open %s: %s", testJsonFile, err)
+		return fmt.Errorf("open %s: %s", testJsonFile, err)
 	}
 	defer f.Close()
 
-	var failedTests strings.Builder
 	for _, jsonFile := range matches {
 		jf, err := os.Open(jsonFile)
 		if err != nil {
-			return "", fmt.Errorf("open %s: %s", jsonFile, err)
+			return fmt.Errorf("open %s: %s", jsonFile, err)
 		}
 
-		var buf bytes.Buffer
-		w := io.MultiWriter(f, &buf)
-		_, err = io.Copy(w, jf)
+		_, err = io.Copy(f, jf)
 		_ = jf.Close()
 		if err != nil {
-			return "", fmt.Errorf("%s copy: %s", jsonFile, err)
-		}
-
-		scanner := bufio.NewScanner(&buf)
-		for scanner.Scan() {
-			var ev testEvent
-			data := scanner.Bytes()
-			if err := json.Unmarshal(data, &ev); err != nil {
-				return "", fmt.Errorf("json unmarshal `%s`: %s", string(data), err)
-			}
-			if ev.Action == "fail" && ev.Test != "" {
-				failedTests.WriteString(fmt.Sprintf("FAIL: %s %s\n", ev.Package, ev.Test))
-			}
-		}
-		if err := scanner.Err(); err != nil {
-			return "", fmt.Errorf("json line scan: %s", err)
+			return fmt.Errorf("%s copy: %s", jsonFile, err)
 		}
 	}
-
-	return failedTests.String(), nil
+	return nil
 }
 
 func testPass() error {
-	var runErrors []string
-
 	matches, err := glob(TestDirRoot, Testsuite, func(path string) bool {
 		return true
 	})
@@ -222,21 +188,13 @@ func testPass() error {
 		cmd.Stderr = os.Stderr
 
 		if err := cmd.Run(); err != nil {
-			runErrors = append(runErrors, fmt.Errorf("cmd run %s: %s", file, err).Error())
+			// log but do not return error
+			fmt.Fprintf(os.Stderr, "cmd run %s: %s", file, err)
 		}
 	}
 
-	failedTests, err := concatenateJsons(jsonPath, jsonOutPath)
-	if err != nil {
+	if err := concatenateJsons(jsonPath, jsonOutPath); err != nil {
 		return fmt.Errorf("concat json: %s", err)
-	}
-	if len(failedTests) > 0 {
-		fmt.Fprintf(os.Stderr, "%s\n", color.RedString(failedTests))
-		// do not return an error here, because we need the rest of the CI script to continue regardless of test failure
-		return nil
-	}
-	if len(runErrors) > 0 {
-		return fmt.Errorf("test binaries had non-zero exit code, but there was no failed tests:\n%s", strings.Join(runErrors, "\n"))
 	}
 	return nil
 }
